refactoring by creating packages for differente responsibilities
