+++ dfe-engine/config.json
@@ -8,7 +8,7 @@
             "extension": "xml",
             "verb": "get",
             "limit": "50",
-            "cursor": 400
+            "cursor": 450
         },
         "nfe/events": {
             "endpoint": "v2/nfe/events",
