+++ dfe-engine/requirements.txt
@@ -1,2 +1,7 @@
+certifi==2021.10.8
+chardet==4.0.0
+idna==2.10
+python-decouple==3.5
 requests==2.25.1
-xmltodict~=0.12.0
+urllib3==1.26.7
+xmltodict==0.12.0
